+++ examples/v2/next-pages-router/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @copilotkitnext/next-pages-router
 
+## 1.53.1-next.2
+
+### Patch Changes
+
+- Updated dependencies [d746e4f]
+  - @copilotkitnext/react@1.53.1-next.2
+
+## 1.53.1-next.1
+
+### Patch Changes
+
+- @copilotkitnext/react@1.53.1-next.1
+
 ## 1.53.1-next.0
 
 ### Patch Changes