+++ packages/v1/a2ui-renderer/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @copilotkit/a2ui-renderer
 
+## 1.53.1-next.2
+
+## 1.53.1-next.1
+
 ## 1.53.1-next.0
 
 ### Patch Changes