+++ packages/v1/react-core/CHANGELOG.md
@@ -1,5 +1,24 @@
 # ui
 
+## 1.53.1-next.2
+
+### Patch Changes
+
+- Updated dependencies [d746e4f]
+  - @copilotkitnext/react@1.53.1-next.2
+  - @copilotkit/runtime-client-gql@1.53.1-next.2
+  - @copilotkit/shared@1.53.1-next.2
+  - @copilotkitnext/core@1.53.1-next.2
+
+## 1.53.1-next.1
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.53.1-next.1
+- @copilotkit/shared@1.53.1-next.1
+- @copilotkitnext/core@1.53.1-next.1
+- @copilotkitnext/react@1.53.1-next.1
+
 ## 1.53.1-next.0
 
 ### Patch Changes