+++ packages/v1/react-ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # ui
 
+## 1.53.1-next.2
+
+### Patch Changes
+
+- @copilotkit/react-core@1.53.1-next.2
+- @copilotkit/runtime-client-gql@1.53.1-next.2
+- @copilotkit/shared@1.53.1-next.2
+
+## 1.53.1-next.1
+
+### Patch Changes
+
+- @copilotkit/react-core@1.53.1-next.1
+- @copilotkit/runtime-client-gql@1.53.1-next.1
+- @copilotkit/shared@1.53.1-next.1
+
 ## 1.53.1-next.0
 
 ### Patch Changes