+++ packages/v2/eslint-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @copilotkitnext/eslint-config
 
+## 1.53.1-next.2
+
+## 1.53.1-next.1
+
 ## 1.53.1-next.0
 
 ## 1.53.0